Added Increased Functionality and Bug Fixes
Added Functions to lock wheels in place to shoot without dislodging
the robot. LockIt() sets stopping settings to hold, UnlockIt() reverses
this. Bounded these to Up and Down buttons. Removed placeholder
autonomous code and replaced with a 10 second shoot. Removed
unnecessary logic in triggerHappy() and replaced it with a simpler one.

Quick Note: We may need to change lines 151 and 156 to be both backward
for ease of driver use.